refactor(api): migrate lookup cache to a static cache

+++ src/api/Features/Geocoding/PoBoxAddressCorrection.cs
@@ -1,6 +1,6 @@
 namespace AGRC.api.Features.Geocoding {
     public class PoBoxAddressCorrection : PoBoxAddress {
-        public PoBoxAddressCorrection(int zip, int zip9, decimal x, decimal y) : base(zip, x, y) {
+        public PoBoxAddressCorrection(int zip, int zip9, double x, double y) : base(zip, x, y) {
             ZipPlusFour = zip9;
         }
